test futures and async code
very similar to trait objects, since Future is a trait. The tested implementation uses trait object wrapping to expose the future and a corresponding `await` function to C. Memory management was a bit of a pain here and difficult to get right, and the `await` implementation necessarily blocks to run the code.